+++ NhakhoaMyNgoc_Connector/DTOs/IdnItemDto.cs
@@ -9,14 +9,14 @@ namespace NhakhoaMyNgoc_Connector.DTOs
         public int IdnId { get; set; }
 
         public string ProductName { get; set; }
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
         public string Unit { get; set; }
 
         public int Quantity { get; set; }
 
         public int Demand { get; set; }
 
-        public int Price { get; set; }
-        public int Total { get; set; }
+        public long Price { get; set; }
+        public long Total { get; set; }
     }
 }
